Sample/EnvironmentLigths package name for Google Play Store
