+++ samples/environment-lights/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.google.ar.sceneform.samples.environmentlights">
+    package="com.gorisse.thomas.ar.environmentlights">
     <!-- "AR Required" apps must declare minSdkVersion ≥ 24.
          "AR Optional" apps must declare minSdkVersion ≥ 14 -->
     <!-- Always needed for AR. -->
